requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
idna==2.10
lxml==4.6.3
parsel==1.6.0
pycodestyle==2.7.0
python-dotenv==0.17.1
PyYAML==5.4.1
requests==2.25.1
schedule==1.1.0
selectorlib==0.16.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.4
w3lib==1.22.0
